Minor updates to paths and checking for package instalations

+++ Portfolios/Code/11_CreateCRSPPredictors.R
@@ -5,41 +5,52 @@
 crsp = read_fst(paste0(pathDataIntermediate,'m_crsp.fst'))
 
 ### MAKE Mom1m
-temp = crsp %>%
-    select(permno, date, ret) %>%
-    mutate(
-        Mom1m = if_else(is.na(ret), 0, ret)
-        , yyyymm = year(date)*100 + month(date)
-    ) %>%
-    filter(!is.na(Mom1m)) %>%
-    select(permno, yyyymm, Mom1m)
+if (!file.exists(paste0(pathPredictors, 'STreversal.csv'))) {
     
-write_csv(temp, paste0(pathCRSPPredictors, 'Mom1m.csv'))
+    temp = crsp %>%
+        select(permno, date, ret) %>%
+        mutate(
+            Mom1m = if_else(is.na(ret), 0, ret)
+            , yyyymm = year(date)*100 + month(date)
+        ) %>%
+        filter(!is.na(Mom1m)) %>%
+        select(permno, yyyymm, Mom1m)
+    
+    write_csv(temp, paste0(pathPredictors, 'STreversal.csv'))
+    
+}
 
 ### MAKE Price
-temp = crsp %>%
-    select(permno, date, prc) %>%
-    mutate(
-        Price = log(abs(prc))
-        , yyyymm = year(date)*100 + month(date)        
-    ) %>%
-    filter(!is.na(Price)) %>%
-    select(permno, yyyymm, Price)
+if (!file.exists(paste0(pathPredictors, 'Price.csv'))) {
     
-write_csv(temp, paste0(pathCRSPPredictors, 'Price.csv'))
+    temp = crsp %>%
+        select(permno, date, prc) %>%
+        mutate(
+            Price = log(abs(prc))
+            , yyyymm = year(date)*100 + month(date)        
+        ) %>%
+        filter(!is.na(Price)) %>%
+        select(permno, yyyymm, Price)
+    
+    write_csv(temp, paste0(pathPredictors, 'Price.csv'))
+    
+}
 
 ### MAKE Size
 # in daily data, I think you want cfacprc to be safe
-temp = crsp %>%
-    select(permno, date, shrout, prc) %>%
-    filter(shrout > 0) %>%    
-    mutate(
-        Size = log(shrout*abs(prc))
-        , yyyymm = year(date)*100 + month(date)        
-    ) %>%
-    filter(!is.na(Size)) %>%
-    select(permno, yyyymm, Size)
+
+if (!file.exists(paste0(pathPredictors, 'Size.csv'))) {
     
-write_csv(temp, paste0(pathCRSPPredictors, 'Size.csv'))
-
-
+    temp = crsp %>%
+        select(permno, date, shrout, prc) %>%
+        filter(shrout > 0) %>%    
+        mutate(
+            Size = log(shrout*abs(prc))
+            , yyyymm = year(date)*100 + month(date)        
+        ) %>%
+        filter(!is.na(Size)) %>%
+        select(permno, yyyymm, Size)
+    
+    write_csv(temp, paste0(pathCRSPPredictors, 'Size.csv'))
+    
+}
